requirements.txt
Django==2.2.8
djangorestframework==3.10.3
Jinja2==2.10.3
MarkupSafe==1.1.1
pytz==2019.3
sqlparse==0.3.0
